+++ mapl/GEOS.F90
@@ -10,12 +10,13 @@ program geos
    integer :: status
    type(MAPL_GriddedComponentDriver) :: driver
    type(ESMF_GridComp), allocatable :: servers(:)
+   type(ESMF_HConfig) :: config
 
-   call MAPL_Initialize(configFileNameFromArgNum=1, _RC)
+   call MAPL_Initialize(configFileNameFromArgNum=1, app_config=config, _RC)
    call MAPL_CreateServers(servers, _RC)
-   call MAPL_CapCreate(driver, _RC)
+   call MAPL_CapCreate(driver, config=config, _RC)
    call MAPL_RunServers(servers, _RC)
-   call MAPL_CapRun(driver, _RC)
+   call MAPL_CapRun(driver, config=config, _RC)
    call MAPL_Finalize(_RC)
 
 end program geos